% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGremlinsConjoint.R
\docType{data}
\name{cbc.df}
\alias{cbc.df}
\title{Simulated data for the "Gremlins in the Data Model"}
\format{
A data frame with 32000 rows and 10 variables:
\describe{
  \item{resp.id}{A respondent identifier}
  \item{ques}{The question or task number}
  \item{alt}{The choice alternative}
  \item{choice}{An indicator that takes on a value of 1 if the alternative was chosen.  (Default is 0.)}
  \item{brandFord}{A dummy coded variable indicating the brand is Ford}
  \item{brandGM}{A dummy coded variable indicating the brand is GM}
  \item{brandDodge}{A dummy coded variable indicating the brand is Dodge}
  \item{enghyb}{A dummy coded variable indicating the engine is a hybrid}
  \item{engelec}{A dummy coded variable indicating the engine is electric}
  \item{price}{A continuous variable for the relative price of the individual offerings.}
}
}
\usage{
cbc.df
}
\description{
A dataset containing simulated choices from a CBC study where some of the
respondents are information poor or 'Gremlins'.  The data is simulated data
and does not reflect actual preferences.
}
\keyword{datasets}
