% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGremlins.R
\name{estimateGremlinsModel}
\alias{estimateGremlinsModel}
\title{Estimate Gremlin's Model - Hierarchical MNL}
\usage{
estimateGremlinsModel(
  data,
  design,
  R,
  Priors = NULL,
  nSegments = 2,
  nBetaSegments = 2,
  covariates = NULL,
  constraints = NULL,
  segmentCovariates = NULL,
  startingValues = NULL
)
}
\arguments{
\item{data}{A matrix containing the raw data.  The format of this data is a N row matrix.
With the first column the respondnet identifier, followed by the design number, then
Number of Observations indexes for the alternative chosen for each of the tasks.  The
index is an integer with 1 representing the first alternative and n representing the
final alternative.}

\item{design}{A Matrix representing the coded design file.  The design file should be
 formatted as a matrix with number of versions * number of tasks * number of alternatives
 rows and number of paramters + 3 columns.  The first column contains the version number,
 The second columns contains the task number, the third number contains the alternative,
 and the remaining columns contain the coded design.  A generic Sawtooth Software design
 file can be converted to this format using the \code{codeSawtoothDesign} function.
@param Priors A datastructure that contains the priors for to the model.  Can be null indicating
 the use of default priors or must contain a full prior specification.
@param numSegments (Default = 2) The number of segments for the scale factor
@param R The number of repitions in the chain
@param covariates (Optional) A matrix of covariates for the model.  One row per respondent with the
 respondent identified in the first column.
@param constraints (Optional) a vector of length n-param sepecifying the constraints
  to impose on the parameters or NULL.  a 1 indicates the parameter is constrained to be positive
  a -1 constrains to be  negative, and a 0 indicates no constraint.
@return A data structure containing the draws from the complete MCMC chain}
}
\description{
The function estimates the model described in "Gremlin's in the Data: Identifying the
Information Content of Research Subjects" using a hierarchical multinomial logit model
}
