% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gremlinsUtilityFunctions.R
\name{generate_starting_atchade_lambdas}
\alias{generate_starting_atchade_lambdas}
\title{Generate starting values for slope Atchade parameters for tuning runs
Adjust Atch_tau_tune_slopes up or down to target
Adjust Atch_tau_tune_slopes and Atch_tau_tune_lambda_in until acceptrates are about 20-80%
(larger Atch_tau_tune_slopes reduce accept rate)
ncluster is number of Germlins clusters
est_mean_lambda      a ncluster by 1 vector, first element has to be one; needs to be good guess of posterior mean lambda
est_var_lambda       a ncluster by 1 vector, first element has to be one; needs to be good guess of posterior variance lambda}
\usage{
generate_starting_atchade_lambdas(
  Atch_tau_tune_lambda,
  ncluster,
  est_mean_lambda,
  est_var_lambda
)
}
\description{
Generate starting values for slope Atchade parameters for tuning runs
Adjust Atch_tau_tune_slopes up or down to target
Adjust Atch_tau_tune_slopes and Atch_tau_tune_lambda_in until acceptrates are about 20-80%
(larger Atch_tau_tune_slopes reduce accept rate)
ncluster is number of Germlins clusters
est_mean_lambda      a ncluster by 1 vector, first element has to be one; needs to be good guess of posterior mean lambda
est_var_lambda       a ncluster by 1 vector, first element has to be one; needs to be good guess of posterior variance lambda
}
