% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gremlinsUtilityFunctions.R
\name{generate_starting_atchade_slopes}
\alias{generate_starting_atchade_slopes}
\title{Generate starting values for slope Atchade parameters for tuning runs
Adjust tuningFactor until accept rates are between 20-80%
(larger tuningFactor reduce accept rate)}
\usage{
generate_starting_atchade_slopes(slopeBar, tuningFactor, nParams, nUnits)
}
\description{
Generate starting values for slope Atchade parameters for tuning runs
Adjust tuningFactor until accept rates are between 20-80%
(larger tuningFactor reduce accept rate)
}
