% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGremlinsConjoint.R
\docType{package}
\name{gremlins}
\alias{gremlins}
\title{gremlins: A package for estimating the "Gremlins in the Data" model}
\description{
The gremlins package provides the tools and utilities to estimate a the model described in
"Gremlin's in the Data: Identifying the Information Content of Research Subjects" using
conjoint analysis data such as that collected in Sawtooth Software's Lighthouse or Discover
Products.  The packages also contains utility functions for formatting the input data and
extracting the relevant results.
}
