% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGremlinsConjoint.R
\docType{data}
\name{gremlinsEnv}
\alias{gremlinsEnv}
\title{Set global options for the gremlins models.  These options are not expected to be modified by the user
but are extracted from the functions to simplify the coding.}
\format{
An object of class \code{environment} of length 4.
}
\usage{
gremlinsEnv
}
\description{
Set global options for the gremlins models.  These options are not expected to be modified by the user
but are extracted from the functions to simplify the coding.
}
\keyword{datasets}
