% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gremlinsUtilityFunctions.R
\name{multivariateRegression}
\alias{multivariateRegression}
\title{Utility function to compute a multivariate regression.  Based on information from "Bayesian Statistics and Marketing (2005)"}
\usage{
multivariateRegression(data, design, priors)
}
\description{
Utility function to compute a multivariate regression.  Based on information from "Bayesian Statistics and Marketing (2005)"
}
