% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGremlinsConjoint-package.R
\docType{package}
\name{RGremlinsConjoint-package}
\alias{RGremlinsConjoint}
\alias{RGremlinsConjoint-package}
\title{RGremlinsConjoint: Estimate the "Gremlins in the Data" Model for Conjoint Studies}
\description{
The tools and utilities to estimate the model described in "Gremlin's in the Data: Identifying the Information Content of Research Subjects" (Howell et al. (2021) \doi{10.1177/0022243720965930}) using conjoint analysis data such as that collected in Sawtooth Software's 'Lighthouse' or 'Discover' products. Additional utilities are included for formatting the input data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/statuser/RGremlinsConjoint}
  \item Report bugs at \url{https://github.com/statuser/RGremlinsConjoint/issues}
}

}
\author{
\strong{Maintainer}: John Howell \email{john_howell@byu.edu} [copyright holder]

Authors:
\itemize{
  \item Peter Ebbes
  \item John Leichty
}

}
\keyword{internal}
