% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGremlins.R
\name{estimateGremlinsModel}
\alias{estimateGremlinsModel}
\title{Estimate Gremlin's Model - Hierarchical MNL}
\usage{
estimateGremlinsModel(
  data,
  design,
  Priors = NULL,
  R = NULL,
  keepEvery = 1,
  nSegments = 2,
  covariates = NULL,
  constraints = NULL,
  segmentCovariates = NULL,
  startingValues = NULL,
  Atch_mcmc_cnt_in,
  Atch_starting_values_slopes_in,
  Atch_starting_values_lambda_in
)
}
\arguments{
\item{data}{A matrix containing the raw data.
The first column the respondent identifier, followed by the design number, the remaining
columns indicate the choices for the tasks that coincide to the design file.}

\item{design}{A matrix representing the coded (dummy of effects) design file.  The design file should be
 formatted as a matrix with number of versions X number of tasks X number of alternatives
 rows and number of parameters + 3 columns.  The first column contains the version number,
 the second columns contains the task number, the third column contains the alternative,
 and the remaining columns contain the coded design.  A generic Sawtooth Software design
 file can be converted to this format using the \code{codeSawtoothDesign} function.
@param Priors A data structure that contains the priors for to the model.  Can be null indicating
 the use of default priors or must contain a full prior specification.
@param R The number of repetitions in the chain
@param keepEvery saves every keepEvery-th draw for output
@param numSegments (Default = 2) The number of segments for the scale factor
@param covariates (Optional) A matrix of covariates for the model.  One row per respondent with the
 respondent identified in the first column.
@param constraints (Optional) a vector of length n-param specifying the constraints
  to impose on the parameters or NULL.  a 1 indicates the parameter is constrained to be positive
  a -1 constrains to be  negative, and a 0 indicates no constraint.
@param segmentCovariates (Optional) a matrix of covariates that influence the probability that each
  individual belongs to the gremlins or reference group.  One row per respondent with the respondent
  identified in the first column.
@param startingValues (Optional) starting values to use for the MCMC algorithm.  This is a list of
  containing: slope = a nRespondent by nParamter matrix of slopes for the respondent
              slopeBar = a nParameter vector of the slopeBar parameter
              slopeCov = a nParameter by nParameter matrix containing the variance covariance matrix
                         for the slopeBar parameter
              lambda = a nSegment vector containing the starting values for the lambda parameters.  The first
                       element in the vector should be 1.
              segMembership = a nRespondent vector containing the segment membership for each respondent.
              phi_lambda = a nParameter vector containing the base probabilities that an individual belongs
                           to each segment.  Should sum to 1.
@param Atch_mcmc_cnt_in ...
@param Atch_starting_values_slopes_in ...
@param Atch_starting_values_lambda_in
@return A data structure containing the draws from the complete MCMC chain}
}
\description{
The function estimates the model described in "Gremlin's in the Data: Identifying the
Information Content of Research Subjects" using a hierarchical multinomial logit model
}
