% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGremlinsConjoint.R
\docType{package}
\name{gremlins}
\alias{gremlins}
\title{'RGremlinsConjoint': A package for estimating the "Gremlins in the Data" model}
\description{
The tools and utilities to estimate the model described in "Gremlin's in the
Data: Identifying the Information Content of Research Subjects" (Howell et al.
(2021) <doi:10.1177/0022243720965930>)  using conjoint analysis data such as
that collected in Sawtooth Software's 'Lighthouse' or 'Discover' products.
Additional utilities are included for formatting the input data.
}
